+++ src/main.cpp
@@ -8,7 +8,7 @@ int main(int argc, char** argv)
     QApplication app(argc, argv);
     QCoreApplication::setOrganizationName(QStringLiteral("ArenaMP"));
     QCoreApplication::setApplicationName(QStringLiteral("ArenaDDSOptimizer"));
-    QCoreApplication::setApplicationVersion(QStringLiteral("0.1.0"));
+    QCoreApplication::setApplicationVersion(QStringLiteral("0.1.3"));
 
     MainWindow w;
     w.show();